social_media_api/requirements.txt
asgiref==3.11.0
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
pillow==10.4.0
PyJWT==2.10.1
sqlparse==0.5.4
